Create /native entry point and move rn specific code to it
