+++ packages/jest/native/index.js
@@ -1,0 +1,1 @@
+export { matchers } from '../src/rn-matchers'
