+++ packages/jest/native/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "dist/jest.cjs.js",
+  "module": "dist/jest.esm.js",
+  "types": "../types/native.d.ts",
+  "preconstruct": {
+    "source": "."
+  }
+}
